Реализация экрана TrackersViewController без вложенной коллекции
